fixed code

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,3 +1,5 @@
+#include "function_pointers.h"
+
 /**
  * array_iterator - Functino that executes another function
  * given as a parameter on each element of an array
@@ -10,7 +12,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < size; i++)
 	{
